Refactored sign in
